Adjusts Cinemachine camera to minimize jitter
The smoothness of player movement depends on a combination of the player's velocity in the y-axis, the lookahead time, and the soft zone height. The camera will likely need to be adjusted to move only in pixel increments to achieve smooth player movement, but the current configuration works well now. For drops in velocity, the soft zone height will likely have to be lowered, and the opposite would be needed for velocity increases. Unsure if applying force with AddForce to the player will affect the current camera setup in terms of preventing jittering.